将代码进行划分
在头文件中加入namespace解决了加头文件不能编译的问题，将几个函数独立放在另一个源文件中。
