+++ 221801305/src/WordCount.h
@@ -1,0 +1,19 @@
+#include <iostream>
+#include <fstream>
+#include <cstdlib>
+#include <cstring>
+#include <vector>
+#include <algorithm>
+#include <iomanip>
+using namespace std;
+
+int countCharacter(char*);
+int countLine(char*);
+int countWord(char*);
+void sortWord();
+void printResult(char*, int, int, int);
+bool isWord(string);
+bool isExist(string);
+string convertToLower(string);
+
+
